src/main/java/com/hotgazpacho/brewhouse/refdata/resource/WeatherResource.java
package com.hotgazpacho.brewhouse.refdata.resource;

import com.hotgazpacho.brewhouse.refdata.weather.WeatherClient;
import com.hotgazpacho.brewhouse.refdata.weather.WeatherResponse;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/api/weather")
public class WeatherResource {


    private final WeatherClient weatherClient;

    @Autowired
    public WeatherResource(final WeatherClient weatherClient) {
        this.weatherClient = weatherClient;
    }

    @GetMapping
    public String weather() {
        return weatherClient.fetchWeather()
                .map(WeatherResponse::getSummary)
                .orElse("Sorry, I couldn't fetch the weather for you :(");
    }
}
